+++ src/main/java/org/nrg/containers/model/xnat/XnatFile.java
@@ -21,7 +21,8 @@ public class XnatFile extends XnatModelObject {
     private List<String> tags;
     private String format;
     private String content;
-    private java.io.File file;
+    private Integer size;
+    private String checksum;
 
     public XnatFile() {}
 
@@ -31,7 +32,8 @@ public class XnatFile extends XnatModelObject {
                     final String tagsCsv,
                     final String format,
                     final String content,
-                    final File file) {
+                    final Integer size,
+                    final String checksum) {
         if (parentUri == null) {
             log.error("Cannot construct a file URI. Parent URI is null.");
         } else {
@@ -44,7 +46,8 @@ public class XnatFile extends XnatModelObject {
         this.tags = Arrays.asList(tagsCsv.split(","));
         this.format = format;
         this.content = content;
-        this.file = file;
+        this.size = size;
+        this.checksum = checksum;
     }
 
     public Project getProject(final UserI userI) {
@@ -92,12 +95,20 @@ public class XnatFile extends XnatModelObject {
         this.content = content;
     }
 
-    public java.io.File getFile() {
-        return file;
+    public Integer getSize() {
+        return size;
     }
 
-    public void setFile(final java.io.File file) {
-        this.file = file;
+    public void setSize(Integer size) {
+        this.size = size;
+    }
+
+    public String getChecksum() {
+        return checksum;
+    }
+
+    public void setChecksum(String checksum) {
+        this.checksum = checksum;
     }
 
     @Override
@@ -121,12 +132,13 @@ public class XnatFile extends XnatModelObject {
                 Objects.equals(this.tags, that.tags) &&
                 Objects.equals(this.format, that.format) &&
                 Objects.equals(this.content, that.content) &&
-                Objects.equals(this.file, that.file);
+                Objects.equals(this.size, that.size) &&
+                Objects.equals(this.checksum, that.checksum);
     }
 
     @Override
     public int hashCode() {
-        return Objects.hash(super.hashCode(), name, path, tags, format, content, file);
+        return Objects.hash(super.hashCode(), name, path, tags, format, content, size, checksum);
     }
 
     @Override